Move service bus test attribute and utilize Microsoft.Extensions.Configuration
